If the storage device is not initializing, just try everything
Part of what makes this solution performant is we initially just try to
initialize storage devices, by initializing driver subsystems:

--subsystem-match=module --subsystem-match=block
--subsystem-match=virtio --subsystem-match=pci --subsystem-match=nvme

there may be some kinds of devices unaccounted for, if this is true,
after 8 seconds try and initialize all devices. If such devices want to
be performant, ideally the people who encounter this should add the
minimal required to this storage specific udev trigger, so their device
is included in the targeted storage trigger.

Signed-off-by: Eric Curtin <ecurtin@redhat.com>